refactor #222 活動取得関連のリファクタリング
