+++ backend/app/domain/income/amount.py
@@ -1,0 +1,2 @@
+def round_money_amount(amount: float) -> float:
+    return round(amount, 2)
