+++ backend/app/repositories/time_repository.py
@@ -3,6 +3,7 @@ from datetime import date
 from sqlalchemy.orm import Session, SessionTransaction
 from sqlalchemy import func, case, extract
 from typing import Optional
+from app.domain.income.amount import round_money_amount
 
 
 class TimeRepository():
@@ -36,6 +37,13 @@ class TimeRepository():
             db_model.Activity.date <= end_date,
             db_model.Activity.username == username).order_by(
                 db_model.Activity.date).all()
+
+    def count_all_activities(self, username: str, status: Optional[str] = None) -> int:
+        sqlstatement = self.db.query(db_model.Activity).filter(
+            db_model.Activity.username == username)
+        if status is not None:
+            sqlstatement = sqlstatement.filter(db_model.Activity.status == status)
+        return sqlstatement.order_by(db_model.Activity.date).count()
 
     def get_all_activities(self, username: str, status: Optional[str] = None) -> list[db_model.Activity]:
         sqlstatement = self.db.query(db_model.Activity).filter(
@@ -83,12 +91,13 @@ class TimeRepository():
             "fail_days": result.fail_days or 0,
         }
 
-    def get_monthly_activity_summary(self, username: str, start_date: date, end_date: date) -> list:
+    def get_monthly_activity_summary(self, username: str, start_date: date, end_date: date) -> dict[int, dict]:
         rows = self.db.query(
             extract("month", db_model.Activity.date).label("month"),
             func.count().label("activity_count"),
             func.sum(case((db_model.Activity.status == "success", 1), else_=0)).label("success_days"),
             func.sum(case((db_model.Activity.status == "failure", 1), else_=0)).label("fail_days"),
+            func.sum(case((db_model.Activity.status == "pending", 1), else_=0)).label("pending_days"),
             func.coalesce(func.sum(db_model.Activity.bonus), 0).label("bonus"),
             func.coalesce(func.sum(db_model.Activity.penalty), 0).label("penalty"),
         ).filter(
@@ -99,4 +108,13 @@ class TimeRepository():
             extract("month", db_model.Activity.date)
         ).all()
 
-        return rows
+        return {
+            int(row.month): {
+                "success_days": row.success_days or 0,
+                "fail_days": (row.fail_days or 0) + (row.pending_days or 0),
+                "bonus": round_money_amount(row.bonus or 0.0),
+                "penalty": round_money_amount(row.penalty or 0.0),
+                "pay_adjustment": round_money_amount((row.bonus or 0.0) - (row.penalty or 0.0)),
+            }
+            for row in rows
+        }
